Remove explicit trace log events
We can use tags to annotate the span with relevant information. We also
don't want to log out the delivery payloads because those can be huge
which can slow down the application. Also it may contain information
that should not be logged.

Also makes sure that tags are using strings as keys. OpenTracing spec
expects strings, not symbols.