Remove publishing to topic exchange in SendAndWaitResponseProducer
There are two producers:
1. SendAndWaitResponseProducer - Used with `Freddy#deliver_with_response`
2. SendAndForgetProducer - Used with `Freddy#deliver`

The reason why we published two messages when using
SendAndWaitResponseProducer was that there's always `Freddy#respond_to`
which uses the default exchange and there can be 0..n `Freddy#tap_into`
listeners which also listen to that message. In real life we don't
actually use the latter one. There's no real need to listen for RPC
calls.

The two exchanges still exist for SendAndForgetProducer. I wanted to
remove the default exchange from there but it is actually used. We could
change our codebase to migrate from respond_to to tap_into in case
`#deliver` is used but that's a separate effort.