Start a new trace when receiving a broadcast
RPC still uses the same trace, but broadcasts (Freddy#deliver) now
create new traces but we still store the original trace link for
observability.

I previously went against the spec because Zipkin doesn't support links.
But we're likely going to replace Zipkin and it's better to follow the
OpenTelemetry spec.