Avoid putting generated queue names to span names
From the OTEL spec:

> The destination name SHOULD only be used for the span name if it is
known to be of low cardinality (cf. general span name guidelines). This
can be assumed if it is statically derived from application code or
configuration. Wherever possible, the real destination names after
resolving logical or aliased names SHOULD be used. If the destination
name is dynamic, such as a conversation ID or a value obtained from a
Reply-To header, it SHOULD NOT be used for the span name. In these
cases, an artificial destination name that best expresses the
destination, or a generic, static fallback like "(temporary)" for
temporary destinations SHOULD be used instead.

In our case I think `(response queue)` makes most sense.