Replace thread gem with concurrent-ruby gem
This gem depends on depends on gem thread which is licensed under WTFPL
license. That gem is not actively maintained and all of its functionality
is included into a well-known gem concurrent-ruby,
which is distributed under MIT license.

Replacing gem dependency and replacing method calls with
corresponding calls to concurrent-ruby library methods.