Bump version to 2.5.0-rc.1
As only an internal library is changed and changes are
backwards-compatible, updating minor version.

The next step after publishing the gem will be updating
it in multiple applications and performing automated tests on
these applications. Gem version will be updated from RC to stable
in a separate pull request after successful validation in the
applications.