Change default queue durability settings
Non-durable queues are not supported with "quorum" type queues and the
only type that supports it, "classic", is no longer present in RabbitMQ
 >=4.0.

Note that "exclusive" queues (i.e. those without a "group" parameter)
are not affected by this change. That is enforced by a conditional right
before queue declaration. All the "group vs durability" checks are thus
unnecessary.

Part of INF-5292